zjwocr/requirement.txt
torch==1.10.0
torchaudio==0.10.0
torchvision==0.11.1
numpy==1.21.4
opencv-python==4.5.4.60
Pillow==8.4.0
requests==2.25.1